requirements.txt
django 
aiogram
pillow
pyTelegramBotAPI 
celery
django-unfold
telepot